Remove legacy files and consolidate to modular architecture
This commit removes all legacy compatibility wrapper files and updates
import statements throughout the codebase to use the new modular
architecture. This cleanup eliminates code duplication and confusion
about which modules to use.

Changes made:
- Removed 9 legacy wrapper files from src/oboyu/indexer/ and src/oboyu/common/
- Updated import statements in tests, benchmarks, and other modules
- All functionality preserved through the new modular service architecture
- Tests pass with no functionality regressions

Files removed:
- src/oboyu/indexer/database.py → use storage/database_service.py
- src/oboyu/indexer/embedding.py → use models/embedding_service.py
- src/oboyu/indexer/reranker.py → use models/reranker_service.py
- src/oboyu/indexer/tokenizer.py → use models/tokenizer_service.py
- src/oboyu/indexer/schema.py → use storage/schema.py
- src/oboyu/indexer/processor.py → use core/document_processor.py
- src/oboyu/indexer/change_detector.py → use storage/change_detector.py
- src/oboyu/common/onnx_converter.py → use common/onnx/ directory
- src/oboyu/indexer/onnx_converter.py → duplicate, removed

Benefits:
- Clearer architecture with consistent import paths
- Reduced maintenance overhead from duplicate code
- Better alignment with service-oriented design
- Simplified codebase navigation
