--- src/oboyu/indexer/change_detector.py
@@ -1,6 +1,0 @@
-"""Legacy import compatibility for change_detector module."""
-
-# Re-export from the new location
-from oboyu.indexer.storage.change_detector import ChangeResult, FileChangeDetector
-
-__all__ = ["ChangeResult", "FileChangeDetector"]
