--- src/oboyu/indexer/tokenizer.py
@@ -1,16 +1,0 @@
-"""Legacy import compatibility for tokenizer module."""
-
-# Re-export from the new location
-from oboyu.indexer.models.tokenizer_service import (
-    FallbackTokenizer,
-    JapaneseTokenizer,
-    TokenizerService,
-    create_tokenizer,
-)
-
-__all__ = [
-    "FallbackTokenizer",
-    "JapaneseTokenizer",
-    "TokenizerService",
-    "create_tokenizer",
-]
